IUFC-91 refactoring information retrieval from api
