--- base/views/autocomplete/common.py
@@ -1,54 +1,0 @@
-##############################################################################
-#
-#    OSIS stands for Open Student Information System. It's an application
-#    designed to manage the core business of higher education institutions,
-#    such as universities, faculties, institutes and professional schools.
-#    The core business involves the administration of students, teachers,
-#    courses, programs and so on.
-#
-#    Copyright (C) 2015-2018 Université catholique de Louvain (http://www.uclouvain.be)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    A copy of this license - GNU General Public License - is available
-#    at the root of the source code of this program.  If not,
-#    see http://www.gnu.org/licenses/.
-#
-##############################################################################
-import requests
-from django.conf import settings
-
-from base.models.academic_year import current_academic_year
-
-
-def get_list_from_osis(url, name_filter=None):
-    header_to_get = {'Authorization': 'Token ' + settings.OSIS_PORTAL_TOKEN}
-    response = requests.get(
-        url=url,
-        headers=header_to_get,
-        params={
-            'search': name_filter or "",
-            'from_year': current_academic_year().year+1,
-            'to_year': current_academic_year().year+2
-        }
-    )
-    data = response.json()
-    if 'results' in data:
-        data = data['results']
-    return data
-
-
-def get_country_list_from_osis(name_filter=None):
-    return get_list_from_osis(settings.URL_COUNTRY_API, name_filter)
-
-
-def get_training_list_from_osis(name_filter=None):
-    return get_list_from_osis(settings.URL_TRAINING_API, name_filter)
