+++ base/views/autocomplete/country.py
@@ -3,7 +3,8 @@ import json
 from dal import autocomplete
 from django import http
 
-from base.views.autocomplete.common import get_country_list_from_osis
+from base.models.academic_year import current_academic_year
+from base.utils.api_utils import get_country_list_from_osis
 
 
 class CountryAutocomplete(autocomplete.Select2ListView):
@@ -12,6 +13,8 @@ class CountryAutocomplete(autocomplete.Select2ListView):
         return http.HttpResponse(json.dumps({
             'results': [
                 {'id': country['uuid'], 'text': country['name']}
-                for country in get_country_list_from_osis(name_filter=self.q)
+                for country in get_country_list_from_osis(
+                    search=self.q,
+                )['results']
             ]
         }), content_type='application/json')
